+++ packages/@expo/metro-config/src/transform-worker/collect-dependencies.ts
@@ -29,7 +29,7 @@ function nullthrows<T extends object>(x: T | null, message?: string): NonNullabl
   return x;
 }
 
-export type AsyncDependencyType = 'weak' | 'maybeSync' | 'async' | 'prefetch';
+export type AsyncDependencyType = 'weak' | 'maybeSync' | 'async' | 'prefetch' | 'worker';
 
 type AllowOptionalDependenciesWithOptions = {
   exclude: string[];
@@ -162,6 +162,33 @@ function collectDependencies<TAst extends t.File>(
   traverse(
     ast,
     {
+      // Match new Worker() patterns
+      NewExpression(path, state: State): void {
+        if (
+          path.node.callee.type === 'Identifier' &&
+          (path.node.callee.name === 'Worker' || path.node.callee.name === 'SharedWorker')
+        ) {
+          const [firstArg] = path.node.arguments;
+
+          // Match: new Worker(new URL("../path/to/module", import.meta.url))
+          if (
+            firstArg &&
+            firstArg.type === 'NewExpression' &&
+            firstArg.callee.type === 'Identifier' &&
+            firstArg.callee.name === 'URL' &&
+            firstArg.arguments.length > 0 &&
+            firstArg.arguments[0].type === 'StringLiteral'
+          ) {
+            const moduleName = firstArg.arguments[0].value;
+
+            // Get the NodePath of the first argument of `new Worker(new URL())`
+            const urlArgPath = (path.get('arguments')[0].get('arguments') as NodePath[])[0];
+
+            processResolveWorkerCallWithName(moduleName, urlArgPath, state);
+          }
+        }
+      },
+
       CallExpression(path, state: State): void {
         if (visited.has(path.node)) {
           return;
@@ -212,6 +239,21 @@ function collectDependencies<TAst extends t.File>(
           !path.scope.getBinding('require')
         ) {
           processResolveWeakCall(path, state);
+          visited.add(path.node);
+          return;
+        }
+
+        // Match `require.unstable_resolveWorker`
+        if (
+          callee.type === 'MemberExpression' &&
+          callee.object.type === 'Identifier' &&
+          callee.object.name === 'require' &&
+          callee.property.type === 'Identifier' &&
+          callee.property.name === 'unstable_resolveWorker' &&
+          !callee.computed &&
+          !path.scope.getBinding('require')
+        ) {
+          processResolveWorkerCall(path, state);
           visited.add(path.node);
           return;
         }
@@ -416,6 +458,49 @@ function processResolveWeakCall(path: NodePath<CallExpression>, state: State): v
       makeResolveWeakTemplate({
         MODULE_ID: createModuleIDExpression(dependency, state),
       })
+    );
+  }
+}
+
+function processResolveWorkerCall(path: NodePath<any>, state: State): void {
+  const name = getModuleNameFromCallArgs(path);
+  if (name == null) {
+    throw new InvalidRequireCallError(path);
+  }
+  return processResolveWorkerCallWithName(name, path, state);
+}
+
+function processResolveWorkerCallWithName(name: string, path: NodePath<any>, state: State): void {
+  const dependency = registerDependency(
+    state,
+    {
+      name,
+      asyncType: 'worker',
+      optional: isOptionalDependency(name, path, state),
+      exportNames: ['*'],
+    },
+    path
+  );
+
+  if (state.collectOnly !== true) {
+    path.replaceWith(
+      makeResolveTemplate({
+        ASYNC_REQUIRE_MODULE_PATH: nullthrows(state.asyncRequireModulePathStringLiteral),
+        DEPENDENCY_MAP: nullthrows(state.dependencyMapIdentifier),
+        MODULE_ID: createModuleIDExpression(dependency, state),
+      })
+    );
+  } else {
+    // Inject the async require dependency into the dependency map so it's available in the graph during serialization and splitting.
+    registerDependency(
+      state,
+      {
+        name: nullthrows(state.asyncRequireModulePathStringLiteral).value,
+        asyncType: null,
+        optional: false,
+        exportNames: ['*'],
+      },
+      path
     );
   }
 }
@@ -674,6 +759,10 @@ const makeAsyncImportMaybeSyncTemplate = template.expression(`
   require(ASYNC_REQUIRE_MODULE_PATH).unstable_importMaybeSync(MODULE_ID, DEPENDENCY_MAP.paths)
 `);
 
+const makeResolveTemplate = template.expression(`
+  require(ASYNC_REQUIRE_MODULE_PATH).unstable_resolve(MODULE_ID, DEPENDENCY_MAP.paths)
+`);
+
 const makeAsyncImportMaybeSyncTemplateWithName = template.expression(`
   require(ASYNC_REQUIRE_MODULE_PATH).unstable_importMaybeSync(MODULE_ID, DEPENDENCY_MAP.paths, MODULE_NAME)
 `);
@@ -781,6 +870,7 @@ function getKeyForDependency(qualifier: ImportQualifier): string {
       contextParams.mode,
     ].join('\0');
   }
+
   return key;
 }